src-tauri/scripts/requirements.txt
# Runtime dependencies for TransNetV2 shot detection (ONNX mode).
# Install with: pip install -r requirements.txt
#
# This is all you need -- no PyTorch, no conda, no GPU required.
onnxruntime>=1.17.0
numpy>=1.24.0
ffmpeg-python>=0.2.0
